# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hệ thống Chấm công QR API
  description: |
    API for QR Code Attendance System with GPS verification.

    ## Authentication
    All endpoints require JWT Bearer token except `/auth/login`.

    ## Rate Limiting
    - Authentication: 5 requests/minute
    - Attendance: 10 requests/minute
    - General: 100 requests/minute
  version: 1.0.0
  contact:
    name: API Support
    email: support@company.com
servers:
  - url: http://localhost:8080/api/v1
    description: Development
  - url: https://attendance.company.com/api/v1
    description: Production
...
paths:
...
  /leave-requests:
    post:
      tags: [Leave Request]
      operationId: createLeaveRequest
      summary: Create a leave request
      description: Employee creates a leave request with date, reason, and optional attachments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveRequestCreateRequest'
      responses:
        '201':
          description: Leave request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequestResponse'

    get:
      tags: [Leave Request]
      operationId: listLeaveRequests
      summary: List leave requests
      description: |
        List leave requests with filters.
        EMPLOYEE: only own requests.
        ADMIN/HR_MANAGER: all requests.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
        - name: userId
          in: query
          schema:
            type: integer
          description: Filter by user (ADMIN/HR only)
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: search
          in: query
          schema:
            type: string
          description: Search by user name or reason
      responses:
        '200':
          description: Paginated leave requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequestPageResponse'
  /leave-requests/{id}:
    get:
      tags: [Leave Request]
      operationId: getLeaveRequest
      summary: Get leave request detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Leave request detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequestResponse'

    put:
      tags: [Leave Request]
      operationId: updateLeaveRequest
      summary: Update leave request (only PENDING status)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveRequestUpdateRequest'
      responses:
        '200':
          description: Leave request updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequestResponse'

    delete:
      tags: [Leave Request]
      operationId: deleteLeaveRequest
      summary: Delete/cancel leave request (only PENDING status)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Leave request deleted
  /leave-requests/{id}/approve:
    post:
      tags: [Leave Request]
      operationId: approveLeaveRequest
      summary: Approve a leave request
      description: |
        Admin/HR approves a leave request.
        Automatically creates an attendance record with status XIN_NGHI for that date.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  description: Optional note from approver
      responses:
        '200':
          description: Leave request approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequestResponse'
  /leave-requests/{id}/reject:
    post:
      tags: [Leave Request]
      operationId: rejectLeaveRequest
      summary: Reject a leave request
      description: Admin/HR rejects a leave request with a reason
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveRequestRejectRequest'
      responses:
        '200':
          description: Leave request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequestResponse'
  /leave-requests/upload:
    post:
      tags: [Leave Request]
      operationId: uploadLeaveAttachment
      summary: Upload attachment for leave request
      description: Upload image (JPG/PNG) or PDF file, max 5MB per file, max 3 files
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload (JPG, PNG, or PDF, max 5MB)
      responses:
        '200':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
...